+++ airframe-spec/src/main/scala/wvlet/airframe/spec/AirSpecFramework.scala
@@ -11,15 +11,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package wvlet.airframe.spec.runner
+package wvlet.airframe.spec
 
 import sbt.testing
 import sbt.testing.{Fingerprint, SubclassFingerprint}
+import wvlet.airframe.spec.runner.AirSpecRunner
 import wvlet.log.LogSupport
 
 /**
   * Include this class to your build.sbt:
-  * testFrameworks += new TestFramework("wvlet.airframe.spec.runner.AirSpecFramework")
+  * testFrameworks += new TestFramework("wvlet.airframe.spec.AirSpecFramework")
   */
 class AirSpecFramework extends sbt.testing.Framework with LogSupport {
   import AirSpecFramework._
