+++ airframe-spec/src/main/scala/wvlet/airframe/spec/CompatApi.scala
@@ -19,8 +19,8 @@ import wvlet.airframe.surface.MethodSurface
   * An interface for compatibility between Scala JVM and Scala.js
   */
 private[spec] trait CompatApi {
-  def findCompanionObjectOf(fullyQualifiedName: String, classLoader: ClassLoader): Option[Any]
-  def newInstanceOf(fullyQualifiedName: String, classLoader: ClassLoader): Option[Any]
+  private[spec] def findCompanionObjectOf(fullyQualifiedName: String, classLoader: ClassLoader): Option[Any]
+  private[spec] def newInstanceOf(fullyQualifiedName: String, classLoader: ClassLoader): Option[Any]
   private[spec] def withLogScanner[U](block: => U): U
   private[spec] def findCause(e: Throwable): Throwable
